% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_bologna}
\alias{data_bologna}
\title{Bologna data.}
\format{A \code{matrix} with 120 rows and 96 columns. Values are either 0 (non-urban) or 1 (urban).}
\source{
EEA (2011). Corine land cover 2000 raster data. Technical Report, downloadable at
http://www.eea.europa.eu/data-and-maps/ data/corine-land-cover-2000-raster-1
}
\usage{
data_bologna
}
\description{
A lattice dataset with Bologna Urban Morphological Zones.
}
\details{
This raster/pixel/lattice dataset comes from the EU CORINE Land Cover project (EEA, 2011) and is dated 2011.
It is the result of classifying the original
land cover data into urbanised and non-urbanised zones, known as 'Urban
Morphological Zones' (UMZ, see EEA, 2011). UMZ data are useful to identify shapes
and patterns of urban areas, and thus to detect what is known as urban sprawl.
The Bologna metropolitan area is extracted from the European dataset and is
composed by the municipality of Bologna and the surrounding municipalities.
The dataset is made of 120x96 pixels of size 250x250 metres.
}
\keyword{datasets}
